[PB-1638] refactor BreadcrumbsItem
